Updated and refactored ability to use a Sidebar.tsx
Created Sidebar.tsx which now is the base logic for the any bar coming in from the left or right. It uses custom hook useDrawerState.ts to handle render logic.

It then takes in any half or full prop to handle what is being rendered

Example in NavigationSidebar.tsx which is a pure link and active navigation method.